+++ NEWS.md
@@ -6,7 +6,7 @@
 
 ## Changed
 
-...
+* `plot_1_categorical_var`: add option to include category labels as text inside bar segments (TRUE by default)
 
 ## Fixed